requirements.txt
langchain-groq
python-dotenv
groq
langchain
langchain-community
streamlit